Fixing broken test cases.
Since we added custom code in _add_job_files_for_upload that looped
through the steps this triggered reading the job file. test_emr.py in
BuildStreamingStepTestCase is assigning the steps manually in the
_assert_streaming_step function. Since we we're triggering _get_steps
in the setUp and _steps hadn't been assigned yet the test tried to
read a my_job.py file instead of just seeing that _steps existed and
returning the values the test passes in.

In test_step since we added new 'io' values to JarStep we needed
the tests to know that those values would be expected